decoupling the map tap feature from the actual prospecting screen
